renamed package to adhere to naming conventions, added two more tests, expanded readme. Added github actions workflow
