+++ src/main/java/nl/openvalue/customerstatementprocessor/CustomerStatementProcessorApplication.java
@@ -1,4 +1,4 @@
-package nl.openvalue.CustomerStatementProcessor;
+package nl.openvalue.customerstatementprocessor;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
@@ -6,8 +6,8 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
 @SpringBootApplication
 public class CustomerStatementProcessorApplication {
 
-	public static void main(String[] args) {
-		SpringApplication.run(CustomerStatementProcessorApplication.class, args);
-	}
+    public static void main(String[] args) {
+        SpringApplication.run(CustomerStatementProcessorApplication.class, args);
+    }
 
 }
