+++ src/main/java/nl/openvalue/customerstatementprocessor/rest/FileProcessingRest.java
@@ -1,9 +1,8 @@
-package nl.openvalue.CustomerStatementProcessor.rest;
+package nl.openvalue.customerstatementprocessor.rest;
 
-import nl.openvalue.CustomerStatementProcessor.processor.FileProcessor;
+import nl.openvalue.customerstatementprocessor.processor.FileProcessor;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
@@ -16,18 +15,17 @@ import org.springframework.web.multipart.MultipartFile;
 @RestController
 @RequestMapping("/api/files")
 public class FileProcessingRest {
-    private final Logger logger = LoggerFactory.getLogger(FileProcessingRest.class);
+    private static final Logger LOGGER = LoggerFactory.getLogger(FileProcessingRest.class);
 
     private final FileProcessor fileProcessor;
 
-    @Autowired
     public FileProcessingRest(FileProcessor fileProcessor) {
         this.fileProcessor = fileProcessor;
     }
 
     @PostMapping(value = "/upload/csv", consumes = MediaType.MULTIPART_FORM_DATA_VALUE)
     public ResponseEntity<String> uploadCsv(@RequestParam("file") MultipartFile file) {
-        logger.info("Triggering new file processing via the csv REST endpoint.");
+        LOGGER.info("Triggering new file processing via the csv REST endpoint.");
         if (file.isEmpty()) {
             return ResponseEntity.badRequest().body("CSV file is empty.");
         }
@@ -36,14 +34,14 @@ public class FileProcessingRest {
             fileProcessor.processCsvFile(file);
             return ResponseEntity.ok("CSV file processed successfully.");
         } catch (Exception e) {
-            logger.error("Error processing CSV file", e);
+            LOGGER.error("Error processing CSV file", e);
             return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body("Error processing CSV file.");
         }
     }
 
     @PostMapping(value = "/upload/xml", consumes = MediaType.MULTIPART_FORM_DATA_VALUE)
     public ResponseEntity<String> uploadXml(@RequestParam("file") MultipartFile file) {
-        logger.info("Triggering new file processing via the xml REST endpoint.");
+        LOGGER.info("Triggering new file processing via the xml REST endpoint.");
         if (file.isEmpty()) {
             return ResponseEntity.badRequest().body("XML file is empty.");
         }
@@ -52,7 +50,7 @@ public class FileProcessingRest {
             fileProcessor.processXmlFile(file);
             return ResponseEntity.ok("XML file processed successfully.");
         } catch (Exception e) {
-            logger.error("Error processing XML file", e);
+            LOGGER.error("Error processing XML file", e);
             return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body("Error processing XML file.");
         }
     }
